@@ -1,4 +1,4 @@
-class CreateCheckIns < ActiveRecord::Migration
+class CreateCheckIns < ActiveRecord::Migration[6.0]
   def change
     create_table :check_ins do |t|
       t.integer :user_id
