@@ -1,18 +1,23 @@
 require './config/environment'
 
 class ApplicationController < Sinatra::Base
+  register Sinatra::AuthLane
+  helpers Sinatra::Cookies
 
   configure do
     set :site_name, 'Save Your Ocean'
+    set :page_title, ' '
     set :public_folder, 'public'
     set :views, 'app/views'
     set :error_msg, 'Sorry'
     enable :sessions
-    set :session_secret, ENV.fetch('SESSION_SECRET') { SecureRandom.hex(64) }
+    set :session_secret, 'sdf098d9jsa'
   end
+  
 
   get "/" do
-    @page_title = "Welcome"
+    @user = current_user
+    settings.page_title = "Welcome to Save Your Ocean!"
     erb :welcome
   end
 
@@ -29,8 +34,7 @@ class ApplicationController < Sinatra::Base
     end
 
     def current_user
-      @current_user = User.find_by(id: session[:user_id])
+      @current_user ||= User.find_by(id: session[:user_id])
     end
-
   end
 end
