@@ -11,8 +11,16 @@ class LocationsController < ApplicationController
   end
 
   # POST: /locations
-  post "/locations" do
-    redirect "/locations"
+  post "/locations/new" do
+    @user = current_user
+    if logged_in?
+      @location = Location.create(params)
+      LocationManager.create(user_id: @user.id, location_id: @location.id)
+
+      redirect to '/users/dashboard'
+    else
+      erb :'/errors/login'
+    end
   end
 
   # GET: /locations/5
