@@ -7,6 +7,6 @@ ActiveRecord::Base.establish_connection(
   :adapter => "sqlite3",
   :database => "db/development.sqlite"
 )
-
+require 'sinatra/authlane'
 require './app/controllers/application_controller'
 require_all 'app'
