@@ -10,7 +10,15 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_05_30_023615) do
+ActiveRecord::Schema.define(version: 2020_06_01_000941) do
+
+  create_table "check_ins", force: :cascade do |t|
+    t.integer "user_id"
+    t.integer "location_id"
+    t.integer "count"
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
+  end
 
   create_table "location_managers", force: :cascade do |t|
     t.integer "location_id"
@@ -29,6 +37,13 @@ ActiveRecord::Schema.define(version: 2020_05_30_023615) do
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.boolean "approved"
+    t.string "zipcode"
+  end
+
+  create_table "roles", force: :cascade do |t|
+    t.string "title"
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
   end
 
   create_table "users", force: :cascade do |t|
@@ -38,6 +53,7 @@ ActiveRecord::Schema.define(version: 2020_05_30_023615) do
     t.string "bio"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.integer "role_id"
   end
 
 end
